Extracted gamelist methods to a different store
