+++ src/stores/game.ts
@@ -7,27 +7,24 @@ import { useUserStore } from "./user"
 export const useGameStore = defineStore('game', () => {
   const userStore = useUserStore();
   const db = getDatabase();
-
+  
   const offGameFuncs: { (): void; }[] = [];
   const offWhitePlayersFuncs: { (): void; }[] = [];
   const offGameListFuncs: { (): void; }[] = [];
-
+  
   const gameId: Ref<string | null> = ref(null)
   const admin: Ref<string | null> = ref(null)
   const gameStarted: Ref<boolean> = ref(false)
   const word: Ref<string | null> = ref(null)
   const players: Ref<{ [uid: string]: string }> = ref({})
   const whitePlayers: Ref<{ [uid: string]: boolean }> = ref({})
-
-  const gameList: Ref<string[]> = ref([])
-
+  
   const amIAdmin = computed(() => admin.value == userStore.user?.uid)
-  const isEmptyGameList = computed(() => gameList.value.length == 0)
-
+  
   const generateRandomGameCode = async () => {
     const gameHeadersRef = fRef(db, 'gameList');
     const gameData = (await get(gameHeadersRef)).val();
-
+    
     let gameCode: string = generate4DigitRandomNumber();
     if (gameData !== null) {
       const existingGames = Object.keys(gameData);
@@ -37,7 +34,7 @@ export const useGameStore = defineStore('game', () => {
     }
     return gameCode;
   }
-
+  
   async function createGame() {
     const gameHeadersRef = fRef(db, 'gameList');
     let retry = false;
@@ -50,7 +47,7 @@ export const useGameStore = defineStore('game', () => {
         retry = true;
       }
     } while (retry);
-
+    
     const gameObject = {
       public: {
         admin: userStore.user?.uid,
@@ -63,25 +60,10 @@ export const useGameStore = defineStore('game', () => {
     set(gameDataRef, gameObject);
     return gameCode;
   }
-
-  function deleteGame(gameIndex: number | string) {
-    let gameCode: string;
-    if (typeof gameIndex === 'number') {
-      gameCode = gameList.value[gameIndex];
-    } else {
-      gameCode = gameIndex;
-    }
-
-    const rootRef = fRef(db);
-    const updates: { [path: string]: any } = {};
-    updates[`gameList/${gameCode}`] = null;
-    updates[`gameData/${gameCode}`] = null;
-    update(rootRef, updates);
-  }
-
+  
   async function joinGame(id: string) {
     const gameRef = fRef(db, `gameData/${id}/public`);
-
+    
     const isGameStarted = (await get(child(gameRef, 'gameStarted'))).val();
     // Racing condition if game starts right here or admin is changed.
     // Purpousefully ignored as it is not game-breaker if a player joins an already started game and it should not change admin while the game is started
@@ -95,26 +77,26 @@ export const useGameStore = defineStore('game', () => {
     }
     return false;
   }
-
+  
   function leaveGame(userId?: string) {
     userId = userId || userStore.user?.uid;
-
+    
     const gameRef = fRef(db, `gameData/${gameId.value!}`);
-
+    
     const updates: { [path: string]: any } = {};
     updates[`public/participants/${userId}`] = null;
     updates[`whitePlayers/${userId}`] = null;
     update(gameRef, updates);
-
+    
     if (userId === userStore.user?.uid) {
       detachGame();
     }
   }
-
+  
   // Transfer administrator privileges from current user to another user
   function makeAdmin(userId: string) {
     const rootRef = fRef(db);
-
+    
     const updates: {[path: string]: any} = {};
     updates[`gameData/${gameId.value}/public/admin`] = userId;
     updates[`gameList/${gameId.value}`] = userId;
@@ -123,32 +105,32 @@ export const useGameStore = defineStore('game', () => {
     updates[`gameData/${gameId.value}/whitePlayers/${userId}`] = null;
     update(rootRef, updates);
   }
-
+  
   function startGame(word: string) {
     const gameRef = fRef(db, `gameData/${gameId.value!}`);
-
+    
     const updates: { [path: string]: any } = {};
     updates['public/gameStarted'] = true;
     updates['word'] = word;
     update(gameRef, updates);
   }
-
+  
   function endGame() {
     const gameRef = fRef(db, `gameData/${gameId.value!}`);
-
+    
     const updates: { [path: string]: any } = {};
     updates['public/gameStarted'] = false;
     updates['word'] = '';
     update(gameRef, updates);
   }
-
+  
   function toggleWhitePlayer(userId: string) {
     const whitePlayersRef = fRef(db, `gameData/${gameId.value!}/whitePlayers`);
     runTransaction(child(whitePlayersRef, userId), (currentData) => {
       return currentData ? null : true;
     });
   }
-
+  
   function attachGame(id: string) {
     gameId.value = id;
     
@@ -158,10 +140,10 @@ export const useGameStore = defineStore('game', () => {
     const gameStartedRef = child(gameRef, `public/gameStarted`);
     const gameAdminRef = child(gameRef, `public/admin`);
     const gameWordRef = child(gameRef, `word`);
-
+    
     offGameFuncs.push(onValue(gameAdminRef, (snapshot) => {
       admin.value = snapshot.val();
-
+      
       whitePlayers.value = {};
       offWhitePlayersFuncs.forEach((offFunc) => offFunc());
       offWhitePlayersFuncs.length = 0;
@@ -179,7 +161,7 @@ export const useGameStore = defineStore('game', () => {
         }));
       }
     }));
-
+    
     offGameFuncs.push(onValue(gameStartedRef, (snapshot) => {
       gameStarted.value = snapshot.val();
     }));
@@ -196,7 +178,7 @@ export const useGameStore = defineStore('game', () => {
       }
     }));
   }
-
+  
   function detachGame() {
     offGameFuncs.forEach((offFunc) => offFunc());
     offGameFuncs.length = 0;
@@ -210,31 +192,6 @@ export const useGameStore = defineStore('game', () => {
     players.value = {};
     whitePlayers.value = {};
   }
-
-  function attachGameList() {
-    const gameHeadersRef = fRef(db, 'gameList');
-
-    offGameListFuncs.push(onChildAdded(gameHeadersRef, (snapshot) => {
-      if (snapshot.val() === userStore.user?.uid) {
-        gameList.value.push(snapshot.key!);
-      }
-    }));
-      
-    offGameListFuncs.push(onChildRemoved(gameHeadersRef, (snapshot) => {
-      const index = gameList.value.indexOf(snapshot.key!);
-      if (index > -1) {
-        gameList.value.splice(index, 1);
-      }
-    }));
-  }
-
-  function detachGameList() {
-    offGameListFuncs.forEach((offFunc) => offFunc());
-    offGameListFuncs.length = 0;
-
-    gameList.value = [];
-  }
-
-  return { gameId, admin, gameStarted, word, players, whitePlayers, gameList, amIAdmin, isEmptyGameList, attachGame, detachGame, attachGameList, detachGameList, createGame, deleteGame, startGame, endGame, makeAdmin, toggleWhitePlayer, joinGame, leaveGame }
-})
   
+  return { gameId, admin, gameStarted, word, players, whitePlayers, amIAdmin, attachGame, detachGame, createGame, startGame, endGame, makeAdmin, toggleWhitePlayer, joinGame, leaveGame }
+})
